Incorporate server version in the dimNeedsUpdate$ stream
