+++ src/app/rx-operators.ts
@@ -1,6 +1,8 @@
 // Spelling out each operator we need helps save on bundle size, but it's a pain
 // to type everywhere. So we keep them all here and you just have to import rx-operators.
 import 'rxjs/add/observable/defer';
+import 'rxjs/add/observable/of';
+import 'rxjs/add/observable/combineLatest';
 import 'rxjs/add/observable/timer';
 import 'rxjs/add/observable/fromEventPattern';
 import 'rxjs/add/observable/fromPromise';
@@ -16,4 +18,3 @@ import 'rxjs/add/operator/take';
 import 'rxjs/add/operator/subscribeOn';
 import 'rxjs/add/operator/do';
 import 'rxjs/add/operator/filter';
-import 'rxjs/add/operator/combineLatest';
